Shorten the undo offer and drop Upcoming's subtitle
The undo snackbar goes from SnackbarDuration.Long to Short, ten seconds
to four. The long form was chosen because undo is the only way back, and
that reasoning left out how often the offer appears: completing a task
is the most frequent thing anyone does in this app, and ten seconds of a
bar across the bottom of the list after every tick is the same tax
expressive-motion.md refuses to put on the interaction itself. A user
who ticks the wrong task knows at once.

It costs nothing in reach. Material passes either value through
calculateRecommendedTimeoutMillis with containsControls set, so anyone
who has asked the system for more time to act is given it whichever is
named here.

Upcoming loses its subtitle. It counted the tasks below it, which is the
one thing a list of tasks already shows - they are grouped under their
own day headings. The rule is now written down: a subtitle says
something the list cannot, or the screen does not get one. Today's date
is not in the list and its planned total is a sum of it; Inbox's count
is the size of a pile the user is deciding whether to work through now;
Anytime and the Logbook have never had one and should not.

So the bars are not all the same height, and that is correct rather than
drift. Inventing subtitles for the screens with nothing to say, to make
four headers agree, would be adding noise for symmetry.

Also fixes two tests that had been failing since Focus lost its app bar
several commits ago, which I did not catch because I never ran the class
that asserts screen chrome. Focus is now the one destination that passes
no bar title: it shows a single task, the navigation bar already names
it, and the heading is the task itself or the empty state's headline,
which every screen here is checked for anyway.

Verified on the emulator, pinned with ANDROID_SERIAL: 118/118 across the
whole ui.semantics package, full JVM suite green, and the snackbar timed
on device - present at three seconds, gone by six.

+++ app/src/main/java/com/vignesh/focuslist/ui/task/UndoSnackbar.kt
@@ -48,8 +48,20 @@ fun UndoSnackbarEffect(
                 is PendingUndo.Reschedule -> rescheduledMessage
             },
             actionLabel = undoLabel,
-            // Undo is the only way back, so give it the longer reading time.
-            duration = SnackbarDuration.Long
+            // Short, which is four seconds against the long form's ten.
+            //
+            // The long form was chosen because undo is the only way back, and
+            // that reasoning ignored how often this appears. Completing a task
+            // is the most frequent thing anyone does here, and ten seconds of
+            // a bar across the bottom of the list after every tick is the same
+            // tax `expressive-motion.md` refuses to put on the interaction
+            // itself. A user who ticks the wrong task knows at once.
+            //
+            // It costs nothing in reach: Material passes either value through
+            // `calculateRecommendedTimeoutMillis` with `containsControls` set,
+            // so anyone who has asked the system for more time to act is given
+            // it, whichever of the two is named here.
+            duration = SnackbarDuration.Short
         )
 
         when (result) {
